Create 18.Twin Prime no using arrays
A twin prime is a prime number that is either 2 less or 2 more than another prime number—for example, either member of the twin prime pair (41, 43). In other words, a twin prime is a prime that has a prime gap of two. ... However, it is unknown whether there are infinitely many twin primes or there is a largest pair.