Create 20. MersennePrimes
a Mersenne prime is a prime number that is one less than a power of two. That is, it is a prime number of the form Mn = 2n − 1 for some integer n. They are named after Marin Mersenne, a French Minim friar, who studied them in the early 17th century.
2	22-1=3  is prime
3	23-1=7  is prime
5	25-1=31  is prime
7	27-1=127  is prime
11	211-1=2047  is not prime